refactor: rename PayPalCheckout to PayPalCommerce
