+++ assets/src/css/admin/settings.scss
@@ -592,7 +592,7 @@ a.give-delete {
 //--------------------------------------------------------------
 // Paypal setting page
 //--------------------------------------------------------------
-#give-paypal-checkout-account-manager-field-wrap {
+#give-paypal-commerce-account-manager-field-wrap {
 	.hero-section {
 		display: grid;
 		grid-template-columns: repeat(2, auto);
